Break up effects into individual files
